requirements.txt
Click>=7.0
pandas>=0.25.1
pyBigWig>=0.3.17
matplotlib>=3.1.1
scipy>=1.3.2
lmfit>=1.0.0
pysam>=0.15.4
more-itertools>=8.9.0
PyYAML>=6.0
